requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pillow==10.3.0
numpy==1.26.4
onnxruntime-gpu==1.18.0
huggingface-hub==0.23.2
python-multipart==0.0.9
